+++ CHANGELOG.md
@@ -44,3 +44,4 @@ and this project adheres to the
 
 - Find subcommand shows item ID instead of quantity
 - Refined find subcommand shows header and formats columns
+- Container attribute to specify where an item is located